refactor subscription function. create helper fns for next and error to abstract boilerplate

+++ apps/admin/src/app/pages/category/category-list/category-list.component.ts
@@ -1,13 +1,8 @@
 import { Component, OnInit } from '@angular/core'
 import { Router } from '@angular/router'
-import { ApiError, ApiResponse } from '@nera/core'
 import { Category, CategoryId, CategoryService } from '@nera/category'
-import {
-  ConfirmationDialogueService,
-  ToastMessageService,
-  ToastMessageSeverity,
-  ToastMessageSummary
-} from '@nera/ui'
+import { HelperFns } from '@nera/core'
+import { ConfirmationDialogueService } from '@nera/ui'
 
 @Component({
   selector: 'admin-category-list',
@@ -17,15 +12,10 @@ import {
 export class CategoryListComponent implements OnInit {
   constructor(
     private categoryService: CategoryService,
-    private messageService: ToastMessageService,
     private confirmationService: ConfirmationDialogueService,
-    private router: Router
-  ) {
-    this.categoryService = categoryService
-    this.messageService = messageService
-    this.confirmationService = confirmationService
-    this.router = router
-  }
+    private router: Router,
+    private utils: HelperFns
+  ) {}
 
   categories: Category[] = []
 
@@ -41,33 +31,26 @@ export class CategoryListComponent implements OnInit {
   }
 
   getCategories(): void {
-    this.categoryService.getCategories().subscribe((categorySubscription) => {
+    this.categoryService.getCategories().subscribe(categorySubscription => {
       const categories = categorySubscription.categories
       this.categories = categories
     })
   }
 
-  deleteCategory(id: CategoryId): void {
-    this.confirmationService.handleConfirm(() =>
+  updateState(id: CategoryId) {
+    this.categories = this.categories.filter(category => category.id !== id)
+  }
+
+  confirmation(id: CategoryId): () => void {
+    return () => {
       this.categoryService.deleteCategory(id).subscribe({
-        next: ({ message }: ApiResponse<Category>) => {
-          this.getCategories()
-          if (message) {
-            this.messageService.handleToastMessage(
-              ToastMessageSeverity.Success,
-              ToastMessageSummary.Success,
-              message
-            )
-          }
-        },
-        error: ({ error }: { error: ApiError }) => {
-          this.messageService.handleToastMessage(
-            ToastMessageSeverity.Error,
-            ToastMessageSummary.Error,
-            error.message
-          )
-        }
+        next: this.utils.handleSuccess({ setState: this.updateState(id) }),
+        error: this.utils.handleError()
       })
-    )
+    }
+  }
+
+  deleteCategory(id: CategoryId): void {
+    this.confirmationService.handleConfirm(this.confirmation(id))
   }
 }
